chore: refactor package name to com.epfl.drawyourpath
