+++ app/src/main/AndroidManifest.xml
@@ -21,16 +21,13 @@
             android:name="com.google.android.geo.API_KEY"
             android:value="${MAPS_API_KEY}" />
         <activity
+            android:name=".bootcamp.GreetingActivity"
+            android:exported="false" />
+        <activity
+            android:name=".bootcamp.FirebaseBootcampActivity"
+            android:exported="false" />
+        <activity
             android:name=".mainpage.MainActivity"
-            android:exported="true" />
-        <activity
-            android:name=".GreetingActivity"
-            android:exported="false" />
-        <activity
-            android:name=".FirebaseBootcampActivity"
-            android:exported="false" />
-        <activity
-            android:name=".authentication.GoogleAuthBootcampActivity"
             android:exported="false" />
 
         <activity
