+++ app/src/main/java/com/epfl/drawyourpath/database/Database.kt
@@ -1,0 +1,8 @@
+package com.epfl.drawyourpath.database
+
+import android.widget.EditText
+
+abstract class Database {
+    abstract fun get(email: EditText, phoneNumber: EditText)
+    abstract fun set(email: EditText, phoneNumber: EditText)
+}
