+++ app/src/main/java/com/epfl/drawyourpath/login/LoginActions.kt
@@ -1,0 +1,24 @@
+package com.epfl.drawyourpath.login
+
+import android.os.Bundle
+import androidx.fragment.app.activityViewModels
+import androidx.fragment.app.Fragment
+import android.view.View
+import android.widget.Button
+import com.epfl.drawyourpath.R
+
+/**
+ * A simple [Fragment] subclass.
+ * Use the [LoginActions.newInstance] factory method to
+ * create an instance of this fragment.
+ */
+class LoginActions : Fragment(R.layout.fragment_login_actions) {
+    private val viewModel: LoginViewModel by activityViewModels()
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        val registerButton = view.findViewById<Button>(R.id.BT_Register)
+        registerButton?.setOnClickListener { viewModel.showRegisterUI() }
+    }
+}
