+++ app/src/main/java/com/epfl/drawyourpath/mainpage/fragments/CommunityFragment.kt
@@ -1,0 +1,8 @@
+package com.epfl.drawyourpath.mainpage.fragments
+
+import androidx.fragment.app.Fragment
+import com.epfl.drawyourpath.R
+
+class CommunityFragment : Fragment(R.layout.fragment_community) {
+    // TODO: add community UI here (like rating/competition) and the ViewModel for this screen/fragment.
+}
