+++ app/src/main/java/com/epfl/drawyourpath/mainpage/fragments/DrawFragment.kt
@@ -1,0 +1,8 @@
+package com.epfl.drawyourpath.mainpage.fragments
+
+import androidx.fragment.app.Fragment
+import com.epfl.drawyourpath.R
+
+class DrawFragment : Fragment(R.layout.fragment_draw) {
+    // TODO: add Google Map UI here and the ViewModel for this screen/fragment.
+}
