+++ app/src/main/java/com/epfl/drawyourpath/mainpage/fragments/FriendsFragment.kt
@@ -1,0 +1,8 @@
+package com.epfl.drawyourpath.bootcamp.mainpage.fragments
+
+import androidx.fragment.app.Fragment
+import com.epfl.drawyourpath.R
+
+class FriendsFragment : Fragment(R.layout.fragment_friends) {
+    // TODO: add friends UI here and the ViewModel for this screen/fragment.
+}
