+++ app/src/main/java/com/epfl/drawyourpath/mainpage/fragments/HistoryFragment.kt
@@ -1,0 +1,8 @@
+package com.epfl.drawyourpath.mainpage.fragments
+
+import androidx.fragment.app.Fragment
+import com.epfl.drawyourpath.R
+
+class HistoryFragment : Fragment(R.layout.fragment_history) {
+    // TODO: add history UI here and the ViewModel for this screen/fragment.
+}
