+++ app/src/main/java/com/epfl/drawyourpath/mainpage/fragments/ProfileFragment.kt
@@ -1,0 +1,10 @@
+package com.epfl.drawyourpath.mainpage.fragments
+
+import androidx.fragment.app.Fragment
+import com.epfl.drawyourpath.R
+
+class ProfileFragment : Fragment(R.layout.fragment_profile) {
+    // TODO: add profile UI and ViewModel for this screen/fragment
+    //  Leaving Profile (with a button for example) should return to the last fragment,
+    //  usually the one select in the bottom menu
+}
