+++ app/src/main/java/com/epfl/drawyourpath/mainpage/fragments/SettingsFragment.kt
@@ -1,0 +1,8 @@
+package com.epfl.drawyourpath.mainpage.fragments
+
+import androidx.fragment.app.Fragment
+import com.epfl.drawyourpath.R
+
+class SettingsFragment : Fragment(R.layout.fragment_settings) {
+    // TODO: add Settings UI and ViewModel for this screen/fragment
+}
