+++ app/src/main/java/com/epfl/drawyourpath/mainpage/fragments/StatsFragment.kt
@@ -1,0 +1,10 @@
+package com.epfl.drawyourpath.mainpage.fragments
+
+import androidx.fragment.app.Fragment
+import com.epfl.drawyourpath.R
+
+class StatsFragment : Fragment(R.layout.fragment_stats) {
+    // TODO: add Stats UI and ViewModel for this screen/fragment
+    //  Leaving Stats (with a button for example) should return to the last fragment,
+    //  usually the one select in the bottom menu
+}
