--- app/src/main/java/com/github/drawyourpath/bootcamp/Database.kt
@@ -1,8 +1,0 @@
-package com.github.drawyourpath.bootcamp
-
-import android.widget.EditText
-
-abstract class Database {
-    abstract fun get(email: EditText, phoneNumber: EditText)
-    abstract fun set(email: EditText, phoneNumber: EditText)
-}
