--- app/src/main/java/com/github/drawyourpath/bootcamp/authentication/AuthUser.kt
@@ -1,9 +1,0 @@
-package com.github.drawyourpath.bootcamp.authentication
-
-interface AuthUser {
-    fun getEmail(): String
-    fun isAnonymous(): Boolean
-    fun getUid(): String
-    fun getPhoneNumber(): String
-    fun getDisplayName(): String
-}
