--- app/src/main/java/com/github/drawyourpath/bootcamp/mainpage/fragments/CommunityFragment.kt
@@ -1,8 +1,0 @@
-package com.github.drawyourpath.bootcamp.mainpage.fragments
-
-import androidx.fragment.app.Fragment
-import com.github.drawyourpath.bootcamp.R
-
-class CommunityFragment : Fragment(R.layout.fragment_community) {
-    // TODO: add community UI here (like rating/competition) and the ViewModel for this screen/fragment.
-}
