--- app/src/main/java/com/github/drawyourpath/bootcamp/mainpage/fragments/DrawFragment.kt
@@ -1,8 +1,0 @@
-package com.github.drawyourpath.bootcamp.mainpage.fragments
-
-import androidx.fragment.app.Fragment
-import com.github.drawyourpath.bootcamp.R
-
-class DrawFragment : Fragment(R.layout.fragment_draw) {
-    // TODO: add Google Map UI here and the ViewModel for this screen/fragment.
-}
