--- app/src/main/java/com/github/drawyourpath/bootcamp/mainpage/fragments/FriendsFragment.kt
@@ -1,8 +1,0 @@
-package com.github.drawyourpath.bootcamp.mainpage.fragments
-
-import androidx.fragment.app.Fragment
-import com.github.drawyourpath.bootcamp.R
-
-class FriendsFragment : Fragment(R.layout.fragment_friends) {
-    // TODO: add friends UI here and the ViewModel for this screen/fragment.
-}
