--- app/src/main/java/com/github/drawyourpath/bootcamp/mainpage/fragments/HistoryFragment.kt
@@ -1,8 +1,0 @@
-package com.github.drawyourpath.bootcamp.mainpage.fragments
-
-import androidx.fragment.app.Fragment
-import com.github.drawyourpath.bootcamp.R
-
-class HistoryFragment : Fragment(R.layout.fragment_history) {
-    // TODO: add history UI here and the ViewModel for this screen/fragment.
-}
