--- app/src/main/java/com/github/drawyourpath/bootcamp/mainpage/fragments/ProfileFragment.kt
@@ -1,10 +1,0 @@
-package com.github.drawyourpath.bootcamp.mainpage.fragments
-
-import androidx.fragment.app.Fragment
-import com.github.drawyourpath.bootcamp.R
-
-class ProfileFragment : Fragment(R.layout.fragment_profile) {
-    // TODO: add profile UI and ViewModel for this screen/fragment
-    //  Leaving Profile (with a button for example) should return to the last fragment,
-    //  usually the one select in the bottom menu
-}
