--- app/src/main/java/com/github/drawyourpath/bootcamp/mainpage/fragments/SettingsFragment.kt
@@ -1,8 +1,0 @@
-package com.github.drawyourpath.bootcamp.mainpage.fragments
-
-import androidx.fragment.app.Fragment
-import com.github.drawyourpath.bootcamp.R
-
-class SettingsFragment : Fragment(R.layout.fragment_settings) {
-    // TODO: add Settings UI and ViewModel for this screen/fragment
-}
