--- app/src/main/java/com/github/drawyourpath/bootcamp/mainpage/fragments/StatsFragment.kt
@@ -1,10 +1,0 @@
-package com.github.drawyourpath.bootcamp.mainpage.fragments
-
-import androidx.fragment.app.Fragment
-import com.github.drawyourpath.bootcamp.R
-
-class StatsFragment : Fragment(R.layout.fragment_stats) {
-    // TODO: add Stats UI and ViewModel for this screen/fragment
-    //  Leaving Stats (with a button for example) should return to the last fragment,
-    //  usually the one select in the bottom menu
-}
