+++ app/src/main/res/layout/activity_login.xml
@@ -48,7 +48,7 @@
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/fragment_container_view"
-            android:name="com.github.drawyourpath.bootcamp.login.RegisterActions"
+            android:name="com.epfl.drawyourpath.bootcamp.login.RegisterActions"
             android:layout_width="match_parent"
 			android:layout_height="wrap_content"
             tools:layout="@layout/fragment_register_actions" />
